Add test project and some refactoring
