+++ TollFeeCalculator/Motorbike.cs
@@ -1,0 +1,11 @@
+﻿namespace TollFeeCalculator
+{
+    public class Motorbike : Vehicle
+    {
+        public string GetVehicleType()
+        {
+            return "Motorbike";
+
+        }
+    }
+}
