+++ TollFeeCalculator/Program.cs
@@ -1,0 +1,6 @@
+﻿namespace TollFeeCalculator
+{
+    internal class Program
+    {
+    }
+}
